Remove symbol.h
