+++ gcc/d/Make-lang.in
@@ -53,7 +53,7 @@ D_DMD_H := \
 
 D_TREE_H = $(TREE_H) d/d-tree.def d/d-lang.h d/d-codegen.h \
         d/d-objfile.h d/d-irstate.h d/d-dmd-gcc.h d/d-system.h \
-        d/d-gcc-real.h d/symbol.h \
+        d/d-gcc-real.h \
         coretypes.h function.h $(VARRAY_H) \
 	$(SYSTEM_H) $(CONFIG_H) $(TARGET_H) \
 	$(GGC_H) \
