+++ gcc/d/d-objfile.h
@@ -18,6 +18,12 @@
 #ifndef GCC_DCMPLR_OBFILE_H
 #define GCC_DCMPLR_OBFILE_H
 
+#include "config.h"
+#include "coretypes.h"
+
+#include "root.h"
+#include "mtype.h"
+
 // These must match the values in object_.d
 enum ModuleInfoFlags
 {
@@ -32,6 +38,48 @@ enum ModuleInfoFlags
   MIimportedModules = 0x400,
   MIlocalClasses    = 0x800,
   MInew		    = 0x80000000
+};
+
+enum OutputStage
+{
+  NotStarted,
+  InProgress,
+  Finished
+};
+
+struct FuncFrameInfo;
+typedef ArrayBase<struct Thunk> Thunks;
+typedef union tree_node dt_t;
+
+struct Symbol : Object
+{
+  Symbol (void);
+
+  const char *Sident;
+  const char *prettyIdent;
+  int Salignment;
+  bool Sreadonly;
+
+  dt_t *Sdt;
+
+  // Specific to GNU backend
+  tree Stree;
+  tree ScontextDecl; // The DECL_CONTEXT to use for child declarations, but see IRState::declContext
+  tree SframeField;  // FIELD_DECL in frame struct that this variable is allocated in
+
+  // For FuncDeclarations:
+  Thunks thunks;
+  OutputStage outputStage;
+  FuncFrameInfo *frameInfo;
+};
+
+struct Thunk
+{
+  Thunk (void)
+  { offset = 0; symbol = NULL; }
+
+  int offset;
+  Symbol *symbol;
 };
 
 extern dt_t **dt_last (dt_t **pdt);
